% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_paths.R
\name{geih_data_base}
\alias{geih_data_base}
\title{Ruta del bundle embebido (GEIH 2024 de ejemplo)}
\usage{
geih_data_base()
}
\value{
Cadena con la ruta absoluta.
}
\description{
Retorna la ruta absoluta al \emph{bundle} de datos incluido dentro del paquete,
ubicado en \verb{inst/extdata/geih_2024/}.
}
\details{
Usado internamente por \code{\link[=merge_month]{merge_month()}} y otras funciones cuando
no se especifica carpeta del usuario.
}
\keyword{internal}
