% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_funciones_auxiliares.R
\name{get_geih_bundle}
\alias{get_geih_bundle}
\title{Obtener y descomprimir bundle GEIH incluido en el paquete}
\usage{
get_geih_bundle()
}
\value{
Ruta absoluta a la carpeta descomprimida \code{geih_2024}.
}
\description{
Esta función busca el archivo \code{inst/extdata/geih_2024.zip} dentro del
paquete, lo descomprime en una carpeta temporal y retorna la ruta.
}
\keyword{internal}
