% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_geih_data_dir}
\alias{get_geih_data_dir}
\title{Obtener carpeta de datos del usuario}
\usage{
get_geih_data_dir()
}
\value{
Ruta absoluta configurada, o \code{NULL} si no existe.
}
\description{
Devuelve la carpeta configurada previamente con \code{\link[=set_geih_data_dir]{set_geih_data_dir()}}.
}
\keyword{internal}
