% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_funciones_auxiliares.R
\name{get_month_divisor}
\alias{get_month_divisor}
\title{Obtener divisor por meses en función de la data}
\usage{
get_month_divisor(data, year = 2024, months = NULL)
}
\arguments{
\item{data}{data.table con columna \code{MES}}

\item{year}{Año de referencia (default: 2024)}

\item{months}{Vector opcional de meses seleccionados}
}
\value{
Número de meses únicos (para dividir expansores)
}
\description{
Obtener divisor por meses en función de la data
}
