% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all_geih.R
\name{join_all_geih}
\alias{join_all_geih}
\title{Hace un empalme de los modulos y los meses de la GEIH
Los datos disponibles son del año 2024}
\usage{
join_all_geih()
}
\value{
Archivo con la GEIH empleta, para el año 2024
}
\description{
\code{join_all_geih()} Crea un archivo .csv llamado "geih_complete.csv".
}
