% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_join_all_geih.R
\name{join_all_months}
\alias{join_all_months}
\title{Fusionar todos los meses de un a単o de la GEIH}
\usage{
join_all_months(year = 2024, data_dir = NULL, verbose = TRUE)
}
\arguments{
\item{year}{A単o (default 2024)}

\item{data_dir}{Carpeta base (opcional)}

\item{verbose}{Mensajes de progreso}
}
\value{
data.table consolidado del a単o
}
\description{
Fusionar todos los meses de un a単o de la GEIH
}
