% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_funciones_auxiliares.R
\name{map_departamentos}
\alias{map_departamentos}
\title{Reemplazar códigos de departamentos por nombres}
\usage{
map_departamentos(dt)
}
\arguments{
\item{dt}{data.table con columna \code{DPTO}}
}
\value{
data.table con DPTO renombrado
}
\description{
Reemplazar códigos de departamentos por nombres
}
