% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_paths.R
\name{set_geih_data_dir}
\alias{set_geih_data_dir}
\title{Configurar carpeta de datos del usuario}
\usage{
set_geih_data_dir(path)
}
\arguments{
\item{path}{Ruta a la carpeta base (ejemplo: \code{"C:/mi_geih_data"}).}
}
\description{
Permite establecer una carpeta propia donde se encuentren los datos GEIH.
Una vez configurada, esta ruta queda guardada como opción global del paquete
y será utilizada automáticamente por funciones como \code{\link[=merge_month]{merge_month()}}.
}
