% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_join_all_geih.R
\name{join_all_geih}
\alias{join_all_geih}
\title{Consolidar GEIH completa de un año}
\usage{
join_all_geih(year = 2024, data_dir = NULL, verbose = TRUE)
}
\arguments{
\item{year}{Año a consolidar (default 2024).}

\item{data_dir}{Ruta base con los datos (opcional). Si es \code{NULL},
se utiliza el bundle embebido comprimido dentro del paquete.}

\item{verbose}{Si TRUE, muestra mensajes de progreso.}
}
\description{
Consolidar GEIH completa de un año
}
