% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_join_all_geih.R
\name{join_all_geih}
\alias{join_all_geih}
\title{Consolidar GEIH completa de un año}
\usage{
join_all_geih(year = 2024, data_dir = NULL, verbose = TRUE)
}
\arguments{
\item{year}{Año (default 2024)}

\item{data_dir}{Carpeta base (opcional)}

\item{verbose}{Mensajes de progreso}
}
\description{
Consolidar GEIH completa de un año
}
